# src/OpenBanking/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API sets created by the Australian Consumer Data Standards to meet the needs of the Consumer Data Right",
    "version" : "1",
    "title" : "Consumer Data Standards",
    "license" : {
      "name" : "MIT Licence",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "data.provider.com.au",
  "basePath" : "/cds-au/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/banking/products" : {
      "get" : {
        "tags" : [ "Products" ],
        "summary" : "Get Products",
        "description" : "Obtain a list of products that are currently openly offered to the market",
        "operationId" : "listProducts",
        "parameters" : [ {
          "name" : "effective",
          "in" : "query",
          "description" : "Allows for the filtering of products based on whether the current time is within the period of time defined as effective by the effectiveFrom and effectiveTo fields.  If absent defaults to 'CURRENT'",
          "required" : false,
          "type" : "string",
          "default" : "CURRENT",
          "enum" : [ "CURRENT", "FUTURE", "ALL" ]
        }, {
          "name" : "updated-since",
          "in" : "query",
          "description" : "Only include products that have been updated after the specified date and time. If absent defaults to include all products",
          "required" : false,
          "type" : "string",
          "format" : "DateTimeString"
        }, {
          "name" : "brand",
          "in" : "query",
          "description" : "Filter results based on a specific brand",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "product-category",
          "in" : "query",
          "description" : "Used to filter results on the productCategory field applicable to accounts. Any one of the valid values for this field can be supplied. If absent then all accounts returned.",
          "required" : false,
          "type" : "string",
          "enum" : [ "PERS_AT_CALL_DEPOSITS", "BUS_AT_CALL_DEPOSITS", "TERM_DEPOSITS", "RESIDENTIAL_MORTGAGES", "PERS_CRED_AND_CHRG_CARDS", "BUS_CRED_AND_CHRG_CARDS", "PERS_LOANS", "PERS_LEASING", "BUS_LEASING", "TRADE_FINANCE", "PERS_OVERDRAFT", "BUS_OVERDRAFT", "BUS_LOANS", "FOREIGN_CURR_AT_CALL_DEPOSITS", "FOREIGN_CURR_TERM_DEPOSITS", "FOREIGN_CURR_LOAN", "FOREIGN_CURRRENCT_OVERDRAFT", "TRAVEL_CARD" ]
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page of results to request (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 1
        }, {
          "name" : "page-size",
          "in" : "query",
          "description" : "Page size to request. Default is 25 (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 25
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingProducts"
            }
          }
        },
        "x-scopes" : [ ],
        "x-version" : "1"
      }
    },
    "/banking/products/{productId}" : {
      "get" : {
        "tags" : [ "Products" ],
        "summary" : "Get Product Detail",
        "description" : "Obtain detailed information on a single product offered openly to the market",
        "operationId" : "getProductDetail",
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "ID of the specific product requested",
          "required" : true,
          "type" : "string",
          "format" : "ASCIIString"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingProduct"
            }
          }
        },
        "x-scopes" : [ ],
        "x-version" : "1"
      }
    },
...
  },
  "definitions" : {
...
    "ResponseBankingProducts" : {
      "type" : "object",
      "required" : [ "data", "links", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/ResponseBankingProducts_data"
        },
        "links" : {
          "$ref" : "#/definitions/LinksPaginated"
        },
        "meta" : {
          "$ref" : "#/definitions/MetaPaginated"
        }
      },
      "example" : {
        "data" : {
          "products" : [ {
            "effectiveTo" : "effectiveTo",
            "lastUpdated" : "lastUpdated",
            "additionalInformation" : {
              "eligibilityUri" : "eligibilityUri",
              "bundleUri" : "bundleUri",
              "feesAndPricingUri" : "feesAndPricingUri",
              "termsUri" : "termsUri",
              "overviewUri" : "overviewUri"
            },
            "brandName" : "brandName",
            "isTailored" : true,
            "productId" : "productId",
            "name" : "name",
            "description" : "description",
            "applicationUri" : "applicationUri",
            "effectiveFrom" : "effectiveFrom",
            "brand" : "brand",
            "productCategory" : { }
          }, {
            "effectiveTo" : "effectiveTo",
            "lastUpdated" : "lastUpdated",
            "additionalInformation" : {
              "eligibilityUri" : "eligibilityUri",
              "bundleUri" : "bundleUri",
              "feesAndPricingUri" : "feesAndPricingUri",
              "termsUri" : "termsUri",
              "overviewUri" : "overviewUri"
            },
            "brandName" : "brandName",
            "isTailored" : true,
            "productId" : "productId",
            "name" : "name",
            "description" : "description",
            "applicationUri" : "applicationUri",
            "effectiveFrom" : "effectiveFrom",
            "brand" : "brand",
            "productCategory" : { }
          } ]
        },
        "meta" : {
          "totalRecords" : 0,
          "totalPages" : 6
        },
        "links" : {
          "next" : "next",
          "last" : "last",
          "prev" : "prev",
          "self" : "self",
          "first" : "first"
        }
      }
    },
    "Product" : {
      "type" : "object",
      "required" : [ "brand", "description", "isTailored", "lastUpdated", "name", "productCategory", "productId" ],
      "properties" : {
        "productId" : {
          "type" : "string",
          "format" : "ASCIIString",
          "description" : "A provider specific unique identifier for this product. This identifier must be unique to a product but does not otherwise need to adhere to ID permanence guidelines."
        },
        "effectiveFrom" : {
          "type" : "string",
          "format" : "DateTimeString",
          "description" : "The date and time from which this product is effective (ie. is available for origination).  Used to enable the articulation of products to the regime before they are available for customers to originate"
        },
        "effectiveTo" : {
          "type" : "string",
          "format" : "DateTimeString",
          "description" : "The date and time at which this product will be retired and will no longer be offered.  Used to enable the managed deprecation of products"
        },
        "lastUpdated" : {
          "type" : "string",
          "format" : "DateTimeString",
          "description" : "The last date and time that the information for this product was changed (or the creation date for the product if it has never been altered)"
        },
        "productCategory" : {
          "$ref" : "#/definitions/EnumProductCategory"
        },
        "name" : {
          "type" : "string",
          "description" : "The display name of the product"
        },
        "description" : {
          "type" : "string",
          "description" : "A description of the product"
        },
        "brand" : {
          "type" : "string",
          "description" : "A label of the brand for the product. Able to be used for filtering. For data providers with single brands this value is still required"
        },
        "brandName" : {
          "type" : "string",
          "description" : "An optional display name of the brand"
        },
        "applicationUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "A link to the an application web page where this product can be applied for."
        },
        "isTailored" : {
          "type" : "boolean",
          "description" : "Indicates whether the product is specifically tailored to a circumstance.  In this case fees and prices are significantly negotiated depending on context. While all products are open to a degree of tailoring this flag indicates that tailoring is expected and thus that the provision of specific fees and rates is not applicable"
        },
        "additionalInformation" : {
          "$ref" : "#/definitions/Product_additionalInformation"
        }
      },
      "example" : {
        "effectiveTo" : "effectiveTo",
        "lastUpdated" : "lastUpdated",
        "additionalInformation" : {
          "eligibilityUri" : "eligibilityUri",
          "bundleUri" : "bundleUri",
          "feesAndPricingUri" : "feesAndPricingUri",
          "termsUri" : "termsUri",
          "overviewUri" : "overviewUri"
        },
        "brandName" : "brandName",
        "isTailored" : true,
        "productId" : "productId",
        "name" : "name",
        "description" : "description",
        "applicationUri" : "applicationUri",
        "effectiveFrom" : "effectiveFrom",
        "brand" : "brand",
        "productCategory" : { }
      }
    },
    "ResponseBankingProduct" : {
      "type" : "object",
      "required" : [ "data", "links", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/ProductDetail"
        },
        "links" : {
          "$ref" : "#/definitions/Links"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        }
      },
      "example" : {
        "data" : "",
        "meta" : { },
        "links" : {
          "self" : "self"
        }
      }
    },
    "ProductDetail" : {
      "allOf" : [ {
        "$ref" : "#/definitions/Product"
      }, {
        "type" : "object",
        "properties" : {
          "bundles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductBundles"
            }
          },
          "features" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductFeatures"
            }
          },
          "constraints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductConstraints"
            }
          },
          "eligibility" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductEligibility"
            }
          },
          "fees" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductFees"
            }
          },
          "depositRates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductDepositRates"
            }
          },
          "lendingRates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/ProductLendingRates"
            }
          }
        }
      } ]
    },
    "ProductBundles" : {
      "type" : "object",
      "required" : [ "description", "name", "productIds" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "description" : "Name of the bundle"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the bundle"
        },
        "additionalInfoUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Link to a web page with more information on the bundle criteria and benefits"
        },
        "productIds" : {
          "type" : "array",
          "description" : "Array of product IDs for products included in the bundle",
          "items" : {
            "type" : "string",
            "format" : "ASCIIString"
          }
        }
      }
    },
    "ProductFeatures" : {
      "type" : "object",
      "required" : [ "featureType" ],
      "properties" : {
        "featureType" : {
          "type" : "string",
          "description" : "The type of feature described",
          "enum" : [ "CARD_ACCESS", "ADDITIONAL_CARDS", "UNLIMITED_TXNS", "FREE_TXNS", "FREE_TXNS_ALLOWANCE", "LOYALTY_PROGRAM", "OFFSET", "OVERDRAFT", "REDRAW", "INSURANCE", "BALANCE_TRANSFERS", "INTEREST_FREE", "INTEREST_FREE_TRANSFERS", "DIGITAL_WALLET", "DIGITAL_BANKING", "NPP_PAYID", "NPP_ENABLED", "DONATE_INTEREST", "BILL_PAYMENT" ]
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the featureType specified. Whether mandatory or not is dependent on the value of featureType"
        }
      },
      "x-conditional" : [ "additionalValue" ]
    },
    "ProductConstraints" : {
      "type" : "object",
      "required" : [ "constraintType" ],
      "properties" : {
        "constraintType" : {
          "type" : "string",
          "description" : "The type of constraint described.  See the next section for an overview of valid values and their meaning",
          "enum" : [ "MIN_BALANCE", "OPENING_BALANCE", "MAX_LIMIT", "MIN_LIMIT" ]
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the constraintType specified.  Whether mandatory or not is dependent on the value of constraintType"
        }
      },
      "x-conditional" : [ "additionalValue" ]
    },
    "ProductEligibility" : {
      "type" : "object",
      "required" : [ "eligibilityType" ],
      "properties" : {
        "eligibilityType" : {
          "type" : "string",
          "description" : "The type of eligibility criteria described.  See the next section for an overview of valid values and their meaning",
          "enum" : [ "BUSINESS", "PENSION_RECIPIENT", "MIN_AGE", "MAX_AGE", "MIN_INCOME", "MIN_TURNOVER", "STAFF", "STUDENT", "EMPLOYMENT_STATUS", "RESIDENCY_STATUS", "OTHER" ]
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the eligibilityType specified.  Whether mandatory or not is dependent on the value of eligibilityType"
        },
        "additionalInfo" : {
          "type" : "string",
          "description" : "Display text providing more information on the eligibility criteria. Mandatory if the eligibilityType field is set to OTHER"
        },
        "additionalInfoUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Link to a web page with more information on this eligibility criteria"
        }
      },
      "x-conditional" : [ "additionalValue", "additionalInfo" ]
    },
    "ProductFees" : {
      "type" : "object",
      "required" : [ "feeType", "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "description" : "Name of the fee"
        },
        "feeType" : {
          "type" : "string",
          "description" : "The type of fee",
          "enum" : [ "PERIODIC", "TRANSACTION", "ESTABLISHMENT", "EXIT", "OVERDRAW", "MIN_BALANCE", "REDRAW", "CHEQUE_CASH", "CHEQUE_STOP", "CHEQUE_BOOK", "CARD_REPLACE", "PAPER_STATEMENT", "OTHER_EVENT" ]
        },
        "amount" : {
          "type" : "string",
          "format" : "AmountString",
          "description" : "The amount charged for the fee. Assumed to be in AUD. One of amount, balanceRate and transactionRate is mandatory"
        },
        "balanceRate" : {
          "type" : "string",
          "format" : "RateString",
          "description" : "A fee rate calculated based on a proportion of the balance. Assumed to be in AUD. One of amount, balanceRate and transactionRate is mandatory"
        },
        "transactionRate" : {
          "type" : "string",
          "format" : "RateString",
          "description" : "A fee rate calculated based on a proportion of a transaction. Assumed to be in AUD. One of amount, balanceRate and transactionRate is mandatory"
        },
        "currency" : {
          "type" : "string",
          "format" : "CurrencyString",
          "description" : "The currency the fee will be charged in. Assumes AUD if absent"
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the feeType specified. Whether mandatory or not is dependent on the value of feeType"
        },
        "additionalInfo" : {
          "type" : "string",
          "description" : "Display text providing more information on the fee"
        },
        "additionalInfoUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Link to a web page with more information on this fee"
        },
        "discounts" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductDiscounts"
          }
        }
      },
      "x-conditional" : [ "additionalValue", "amount", "balanceRate", "transactionRate" ]
    },
    "ProductDiscounts" : {
      "type" : "object",
      "required" : [ "amount", "description", "discountType" ],
      "properties" : {
        "description" : {
          "type" : "string",
          "description" : "Description of the discount"
        },
        "discountType" : {
          "type" : "string",
          "description" : "The type of discount. See the next section for an overview of valid values and their meaning",
          "enum" : [ "BALANCE", "DEPOSITS", "PAYMENTS", "BUNDLE" ]
        },
        "amount" : {
          "type" : "string",
          "format" : "AmountString",
          "description" : "Value of the discount. Note that the currency of the fee discount is expected to be the same as the currency of the fee itself"
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the discountType specified. Whether mandatory or not is dependent on the value of discountType"
        }
      },
      "x-conditional" : [ "additionalValue" ]
    },
    "ProductDepositRates" : {
      "type" : "object",
      "required" : [ "depositRateType", "rate" ],
      "properties" : {
        "depositRateType" : {
          "type" : "string",
          "description" : "The type of rate (base, bonus, etc). See the next section for an overview of valid values and their meaning",
          "enum" : [ "FIXED", "BONUS", "BUNDLE_BONUS", "VARIABLE", "INTRODUCTORY" ]
        },
        "rate" : {
          "type" : "string",
          "format" : "RateString",
          "description" : "The rate to be applied"
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Generic field containing additional information relevant to the depositRateType specified. Whether mandatory or not is dependent on the value of depositRateType"
        },
        "additionalInfo" : {
          "type" : "string",
          "description" : "Display text providing more information on the fee"
        },
        "additionalInfoUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Link to a web page with more information on this fee"
        }
      },
      "x-conditional" : [ "additionalValue" ]
    },
    "ProductLendingRates" : {
      "type" : "object",
      "required" : [ "lendingRateType", "rate" ],
      "properties" : {
        "lendingRateType" : {
          "type" : "string",
          "description" : "The type of rate (fixed, variable, etc). See the next section for an overview of valid values and their meaning",
          "enum" : [ "FIXED", "INTRODUCTORY", "DISCOUNT", "PENALTY", "BUNDLE_DISCOUNT", "FLOATING", "MARKET_LINKED", "CASH_ADVANCE", "VARIABLE", "COMPARISON" ]
        },
        "rate" : {
          "type" : "string",
          "format" : "RateString",
          "description" : "The rate to be applied"
        },
        "additionalValue" : {
          "type" : "string",
          "description" : "Information relevant to the lendingRateType specified.  Whether mandatory or not is dependent on the Generic field containing additional information relevant to the lendingRateType specified. Whether mandatory or not is dependent on the value of lendingRateType"
        },
        "additionalInfo" : {
          "type" : "string",
          "description" : "Display text providing more information on the fee."
        },
        "additionalInfoUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Link to a web page with more information  on this fee"
        }
      },
      "x-conditional" : [ "additionalValue" ]
    },
...
    "Links" : {
      "type" : "object",
      "properties" : {
        "self" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Fully qualified link to this API call"
        }
      },
      "example" : {
        "self" : "self"
      }
    },
    "Meta" : {
      "type" : "object"
    },
    "LinksPaginated" : {
      "type" : "object",
      "properties" : {
        "self" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Fully qualified link to this API call"
        },
        "first" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first page of this set. Mandatory if this response is not the first page"
        },
        "prev" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first previous page of this set. Mandatory if this response is not the first page"
        },
        "next" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first next page of this set. Mandatory if this response is not the last page"
        },
        "last" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first last page of this set. Mandatory if this response is not the last page"
        }
      },
      "example" : {
        "next" : "next",
        "last" : "last",
        "prev" : "prev",
        "self" : "self",
        "first" : "first"
      },
      "x-conditional" : [ "prev", "next", "first", "last" ]
    },
    "MetaPaginated" : {
      "type" : "object",
      "properties" : {
        "totalRecords" : {
          "type" : "integer",
          "format" : "NaturalNumber",
          "description" : "The total number of records in the full set"
        },
        "totalPages" : {
          "type" : "integer",
          "format" : "NaturalNumber",
          "description" : "The total number of pages in the full set"
        }
      },
      "example" : {
        "totalRecords" : 0,
        "totalPages" : 6
      }
    },
...
    "EnumProductCategory" : {
      "type" : "string",
      "description" : "The list of available product categories for categorising products and accounts",
      "enum" : [ "PERS_AT_CALL_DEPOSITS", "BUS_AT_CALL_DEPOSITS", "TERM_DEPOSITS", "RESIDENTIAL_MORTGAGES", "PERS_CRED_AND_CHRG_CARDS", "BUS_CRED_AND_CHRG_CARDS", "PERS_LOANS", "PERS_LEASING", "BUS_LEASING", "TRADE_FINANCE", "PERS_OVERDRAFT", "BUS_OVERDRAFT", "BUS_LOANS", "FOREIGN_CURR_AT_CALL_DEPOSITS", "FOREIGN_CURR_TERM_DEPOSITS", "FOREIGN_CURR_LOAN", "FOREIGN_CURRRENCT_OVERDRAFT", "TRAVEL_CARD" ],
      "x-enum-string" : true
    },
...
    "ResponseBankingProducts_data" : {
      "required" : [ "products" ],
      "properties" : {
        "products" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Product"
          }
        }
      },
      "example" : {
        "products" : [ {
          "effectiveTo" : "effectiveTo",
          "lastUpdated" : "lastUpdated",
          "additionalInformation" : {
            "eligibilityUri" : "eligibilityUri",
            "bundleUri" : "bundleUri",
            "feesAndPricingUri" : "feesAndPricingUri",
            "termsUri" : "termsUri",
            "overviewUri" : "overviewUri"
          },
          "brandName" : "brandName",
          "isTailored" : true,
          "productId" : "productId",
          "name" : "name",
          "description" : "description",
          "applicationUri" : "applicationUri",
          "effectiveFrom" : "effectiveFrom",
          "brand" : "brand",
          "productCategory" : { }
        }, {
          "effectiveTo" : "effectiveTo",
          "lastUpdated" : "lastUpdated",
          "additionalInformation" : {
            "eligibilityUri" : "eligibilityUri",
            "bundleUri" : "bundleUri",
            "feesAndPricingUri" : "feesAndPricingUri",
            "termsUri" : "termsUri",
            "overviewUri" : "overviewUri"
          },
          "brandName" : "brandName",
          "isTailored" : true,
          "productId" : "productId",
          "name" : "name",
          "description" : "description",
          "applicationUri" : "applicationUri",
          "effectiveFrom" : "effectiveFrom",
          "brand" : "brand",
          "productCategory" : { }
        } ]
      }
    },
    "Product_additionalInformation" : {
      "properties" : {
        "overviewUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "General overview of the product"
        },
        "termsUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Terms and conditions for the product"
        },
        "eligibilityUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Eligibility rules and criteria for the product"
        },
        "feesAndPricingUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Description of fees, pricing, discounts, exemptions and bonuses for the product"
        },
        "bundleUri" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Description of a bundle that this product can be part of"
        }
      },
      "example" : {
        "eligibilityUri" : "eligibilityUri",
        "bundleUri" : "bundleUri",
        "feesAndPricingUri" : "feesAndPricingUri",
        "termsUri" : "termsUri",
        "overviewUri" : "overviewUri"
      }
    },
...
  },
...
}